Add UnsavedWarning to list page too
